Update tools.py
Overhaul of the file. As suggested by Andy we are better off breaking all the DEs down into smaller functions which we can then put together to make up final functions for each of the 5 ODEs. The function dydr was changed to be a bit cleaner but follows the same logic Matthew had! Matthew's original code in dydr is just commented out.